Crate strategy_lib

strategy_lib/src/lib.rs

Structs

MovingAverageCrossStrategy

A simple moving average crossover strategy. This strategy generates buy/sell signals based on the crossing of two SMAs. It also handles position exits based on SMA crossover or contract expiration.

Functions

calculate_signals

C function exported for dynamic loading. Calls the calculate_signals method on the MovingAverageCrossStrategy instance. This function is called by the main application during the backtest loop.

create_strategy

C function exported for dynamic loading. Creates a new instance of the MovingAverageCrossStrategy. This function is called by the main application when loading the strategy library.

destroy_strategy

C function exported for dynamic loading. Destroys an instance of the MovingAverageCrossStrategy. This function is called by the main application when unloading the strategy library.